Split test-repr.py into 2 parts
